experimental/teams-tabs/python/requirements.txt
msrest
botbuilder-integration-aiohttp>=4.11.0
